installatin with requirments

+++ echec_vision/pok/classes/chess_plate.py
@@ -14,8 +14,8 @@ class ChessPlate:
         col_ratio = resized_plate_img.shape[1] / plate_img.shape[1]
         row_ratio = resized_plate_img.shape[0] / plate_img.shape[0]
 
-        resize_col_coords = (col_coords * col_ratio).astype(np.int)
-        resize_row_coords = (row_coords * row_ratio).astype(np.int)
+        resize_col_coords = (col_coords * col_ratio).astype(np.int16)
+        resize_row_coords = (row_coords * row_ratio).astype(np.int16)
 
         self.plate_img = resized_plate_img
         self.col_coords = resize_col_coords
